backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.7.2
google-cloud-texttospeech==2.17.2
google-cloud-firestore==2.16.0
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
httpx==0.27.0
pydantic==2.7.1
python-multipart==0.0.9
gTTS==2.5.1

 
